+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Load properties from those available in class to prevent setting dynamic properties
+
 ## 0.0.2
 - Load data into private properties